Moving EnvironmentManager tests to a new dll
