+++ test/EnvironmentManagerTests/EnvironmentManagerCentennialTests.cpp
@@ -1,26 +1,26 @@
-﻿#include <pch.h>
-#include <EnvironmentManagerCentennialTests.h>
+﻿#include "pch.h"
+#include "EnvironmentManagerCentennialTests.h"
 
 using namespace winrt::Microsoft::ProjectReunion;
 
-namespace ProjectReunionCppTest
+namespace ProjectReunionEnvironmentManagerTests
 {
     void EnvironmentManagerCentennialTests::CentennialTestGetForProcess()
     {
-        EnvironmentManager environmentManager = EnvironmentManager::GetForProcess();
+        EnvironmentManager environmentManager{ EnvironmentManager::GetForProcess() };
         VERIFY_IS_NOT_NULL(environmentManager);
     }
 
     void EnvironmentManagerCentennialTests::CentennialTestGetForUser()
     {
-        EnvironmentManager environmentManager = EnvironmentManager::GetForUser();
+        EnvironmentManager environmentManager{ EnvironmentManager::GetForUser() };
         VERIFY_IS_NOT_NULL(environmentManager);
     }
 
 
     void EnvironmentManagerCentennialTests::CentennialTestGetForMachine()
     {
-        EnvironmentManager environmentManager = EnvironmentManager::GetForMachine();
+        EnvironmentManager environmentManager{ EnvironmentManager::GetForMachine() };
         VERIFY_IS_NOT_NULL(environmentManager);
     }
 }
