+++ test/EnvironmentManagerTests/EnvironmentManagerTests.dll.manifest
