+++ test/EnvironmentManagerTests/EnvironmentManagerWin32Tests.cpp
@@ -1,0 +1,26 @@
+﻿#include "pch.h"
+#include "EnvironmentManagerWin32Tests.h"
+
+using namespace winrt::Microsoft::ProjectReunion;
+
+namespace ProjectReunionEnvironmentManagerTests
+{
+    void EnvironmentManagerWin32Tests::TestGetForProcess()
+    {
+        EnvironmentManager environmentManager{ EnvironmentManager::GetForProcess() };
+        VERIFY_IS_NOT_NULL(environmentManager);
+    }
+
+    void EnvironmentManagerWin32Tests::TestGetForUser()
+    {
+        EnvironmentManager environmentManager{ EnvironmentManager::GetForUser() };
+        VERIFY_IS_NOT_NULL(environmentManager);
+    }
+
+
+    void EnvironmentManagerWin32Tests::TestGetForMachine()
+    {
+        EnvironmentManager environmentManager{ EnvironmentManager::GetForMachine() };
+        VERIFY_IS_NOT_NULL(environmentManager);
+    }
+}
