+++ test/EnvironmentManagerTests/EnvironmentManagerWin32Tests.h
@@ -1,13 +1,13 @@
 ﻿#pragma once
 
-namespace ProjectReunionCppTest
+namespace ProjectReunionEnvironmentManagerTests
 {
     class EnvironmentManagerWin32Tests {
         BEGIN_TEST_CLASS(EnvironmentManagerWin32Tests)
-            TEST_CLASS_PROPERTY(L"ActivationContext", L"CppTest.dll.manifest")
+            TEST_CLASS_PROPERTY(L"ActivationContext", L"EnvironmentManagerTests.dll.manifest")
+            TEST_CLASS_PROPERTY(L"ThreadingModel", L"MTA")
         END_TEST_CLASS()
 
-        TEST_METHOD_SETUP(TestSetup);
         TEST_METHOD(TestGetForProcess);
         TEST_METHOD(TestGetForUser);
         TEST_METHOD(TestGetForMachine);
