+++ test/EnvironmentManagerTests/packages.config
@@ -1,0 +1,8 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Microsoft.VCRTForwarders.140" version="1.0.7" targetFramework="native" />
+  <package id="Microsoft.Windows.CppWinRT" version="2.0.200703.9" targetFramework="native" />
+  <package id="Microsoft.Windows.ImplementationLibrary" version="1.0.200519.2" targetFramework="native" />
+  <package id="Taef.Redist.Wlk" version="10.52.200221001" targetFramework="native" />
+  <package id="Taef.TestAdapter" version="10.57.201103002-develop" targetFramework="native" />
+</packages>
