+++ test/EnvironmentManagerTests/pch.h
@@ -1,0 +1,12 @@
+﻿#ifndef PCH_H
+#define PCH_H
+
+#include "framework.h"
+#include "wil/cppwinrt.h"
+#include "wil/result.h"
+#include "winrt/Microsoft.ProjectReunion.h"
+//#include <winrt/Microsoft.ProjectReunion.h>
+#include <windows.h>
+#include "WexTestClass.h"
+
+#endif
